fix model liter field type

+++ item/models.py
@@ -66,7 +66,7 @@ class Item(models.Model):
     main_price = models.DecimalField(default=0, max_digits=7, decimal_places=2)
     sell_price = models.DecimalField(default=0, max_digits=7, decimal_places=2)
     image_url = models.URLField(max_length = 200, blank=True, null=True)
-    liter = models.IntegerField(default=0)
+    liter = models.DecimalField(default=0, max_digits=7, decimal_places=2)
     created_at = models.DateTimeField(auto_now_add=True)
     updated_at = models.DateTimeField(auto_now=True)
